Reorganize command and parser into packages
